% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial.R
\name{createDbSpatPoints}
\alias{createDbSpatPoints}
\title{Create a duckdb table with spatial points}
\usage{
createDbSpatPoints(data, conn, tbl, overwrite = FALSE, ...)
}
\arguments{
\item{data}{Data to be added to database. Can be a data.frame, tbl,
file path, or {terra} spatVector/spatRaster. Experimental support for
lists of files is supported through {duckdbfs}.}

\item{conn}{Connection to duckdb database.}

\item{tbl}{Name of table to add dbSpatPoints to.}

\item{overwrite}{Overwrite existing table (default = FALSE)}

\item{...}{Additional arguments to be passed}
}
\value{
A duckdb table with spatial points.
}
\description{
Create a duckdb table with spatial points
}
\details{
For full list of supported file types see ST_Read() in duckdb.
}
\examples{
# create in-memory duckdb db
duckdb_conn = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# test data
test_data = data.frame(x = 1:10, y = 1:10, id = 1:10)

write.csv(test_data, "test_data.csv", row.names = FALSE)

# df, tbl
createDbSpatPoints(conn = duckdb_conn,
                   tbl = "test_points",
                   data = test_data,
                   overwrite = TRUE)

# file
createDbSpatPoints(conn = duckdb_conn,
                   tbl = "test_points",
                   data = 'test_data.csv',
                   overwrite = TRUE)

}
\keyword{constructor}
