% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial.R
\name{createDbSpatPolygons}
\alias{createDbSpatPolygons}
\title{Create a duckdb table with spatial polygons}
\usage{
createDbSpatPolygons(data, conn, tbl, overwrite = FALSE, ...)
}
\arguments{
\item{data}{Data to be added. Can be a terra::spatRaster, terra::spatVector, or file path.}

\item{conn}{Connection to duckdb database.}

\item{tbl}{Name of table to add dbSpatPolygons to.}

\item{overwrite}{Overwrite existing table (default = FALSE)}

\item{...}{Additional arguments to be passed.}
}
\value{
A duckdb table with spatial polygons.
}
\description{
Create a duckdb table with spatial polygons
}
\examples{
TODO
}
\keyword{constructor}
