% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dbSpatial.R
\name{.sim_dbSpatial}
\alias{.sim_dbSpatial}
\title{Simulate basic dbSpatial object}
\usage{
.sim_dbSpatial()
}
\value{
A dbSpatial object
}
\description{
Simulate basic dbSpatial object
}
\keyword{internal}
