% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_read.R
\name{.st_read}
\alias{.st_read}
\title{Ingest spatial data from a file as a \code{VIEW} and create a \code{dbSpatial} object}
\usage{
.st_read(
  conn,
  name,
  value,
  x_colName = NULL,
  y_colName = NULL,
  geomName = "geom",
  overwrite = FALSE,
  return = TRUE,
  ...
)
}
\arguments{
\item{conn}{a duckdb connection}

\item{name}{name of the table to be created}

\item{value}{a data.frame or a file path}

\item{x_colName}{name of the column containing x coordinates. default: 'NULL'}

\item{y_colName}{name of the column containing y coordinates. default: 'NULL'}

\item{geomName}{name of the geometry column to be created. default: 'geom'}

\item{overwrite}{logical; if TRUE, overwrite the table if it already exists
default: 'FALSE'}

\item{return}{return dbSpatial object. default: 'FALSE'}

\item{...}{additional arguments to pass to st_read}
}
\value{
\code{dbSpatial} object if \code{return = TRUE} else \code{NULL}
}
\description{
Ingest spatial data from a file as a \code{VIEW} and create a \code{dbSpatial} object
}
\details{
For list of files supported see the documentation below.
\url{https://DuckDB.org/docs/extensions/spatial.html#st_read---read-spatial-value-from-files}
}
\keyword{internal}
