% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSpatial}
\alias{loadSpatial}
\title{Install and/or load duckdb spatial extension}
\usage{
loadSpatial(conn)
}
\arguments{
\item{conn}{duckdb connection}
}
\description{
Install and/or load duckdb spatial extension
}
\examples{
duckdb_conn = DBI::dbConnect(duckdb::duckdb(), ":memory:")
loadSpatial(conn = duckdb_conn)
}
\keyword{duckdb-ext}
