% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial.R
\name{nrow}
\alias{nrow}
\title{Count number of rows in a duckdb table}
\usage{
nrow(tbl)
}
\arguments{
\item{tbl}{name of a table in a duckdb database}
}
\value{
number of rows in tbl (integer)
}
\description{
Count number of rows in a duckdb table
}
\examples{
TODO
}
\keyword{spatial_prop}
