% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_difference.R
\name{st_difference}
\alias{st_difference}
\title{Find difference in \code{geometries} between two dbSpatial objects}
\usage{
st_difference(
  g1,
  g1_geom_colName = "geom",
  g1_cols_keep = "all",
  g2,
  g2_geom_colName = "geom",
  g2_cols_keep = "all",
  overwrite = FALSE,
  output_tblName = "difference_geom"
)
}
\arguments{
\item{g1}{A \code{\link{dbSpatial}} object.}

\item{g1_geom_colName}{\code{character}. The \code{geometry} column name in \code{g1}.
default: "geom"}

\item{g1_cols_keep}{\code{character vector}. The column names in \code{g1} to keep
after the operation. default: 'all' for all columns}

\item{g2}{\code{\link{dbSpatial}} object.}

\item{g2_geom_colName}{\code{character}. The \code{geometry} column name in \code{g2}.
default: "geom"}

\item{g2_cols_keep}{\code{character vector}. The column names in \code{g2} to keep
after the operation. default: 'all' for all columns}

\item{overwrite}{\code{logical}. If TRUE, overwrite an existing table with the
same \code{output_tblName}. default: FALSE}

\item{output_tblName}{\code{character}. The name of the table to store the
resulting data. default: "difference_geom"}
}
\value{
tbl_dbi
}
\description{
Adds column 'st_difference' to the resulting table containing the difference
between the geometries in g1 and g2.
}
\details{
\url{https://postgis.net/docs/ST_Difference.html}
}
\examples{
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

polys <- dbSpatial(conn = con,
                   name = "polys",
                   value = system.file("data", "dummy_polygons.geojson", package = "dbSpatial"),
                   overwrite = TRUE,
                   x_colName = "x",
                   y_colName = "y")

polys2 <- dbSpatial(conn = con,
                   name = "polys2",
                   value = system.file("data", "dummy_polygons.geojson", package = "dbSpatial"),
                   overwrite = TRUE,
                   x_colName = "x",
                   y_colName = "y")

res <- st_difference(g1 = polys, 
                     g2 = polys2,
                     overwrite = TRUE)

res
}
\seealso{
Other spatial_relationships: 
\code{\link{st_contains}()},
\code{\link{st_coveredby}()},
\code{\link{st_covers}()},
\code{\link{st_crosses}()},
\code{\link{st_disjoint}()},
\code{\link{st_distance}()},
\code{\link{st_equals}()},
\code{\link{st_intersects}()},
\code{\link{st_touches}()}
}
\concept{spatial_relationships}
