% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_GeometryType.R
\name{ST_GeometryType}
\alias{ST_GeometryType}
\title{Return geometry type}
\usage{
ST_GeometryType(tbl, geomName = "geom", limit = 10)
}
\arguments{
\item{tbl}{name of a table in a duckdb database}

\item{geomName}{name of the column containing the geometry value in the tbl.
default = "geom".}

\item{limit}{number of rows to return. default = 10.}
}
\value{
tbl_dbi
}
\description{
This function returns the geometry type of the specified geometry column in
the specified table.
}
\examples{
# Create a data.frame with x and y coordinates and attributes
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a duckdb connection
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = con,
                      value = dummy_data,
                      x_colName = "x",
                      y_colName = "y",
                      name = "foo",
                      overwrite = TRUE)
                      
ST_GeometryType(tbl = db_points)
}
\keyword{geo_properties}
