% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/st_isvalid.R
\name{st_isvalid}
\alias{st_isvalid}
\alias{st_isvalid,dbSpatial-method}
\title{Determine if geometry is valid}
\usage{
st_isvalid(dbSpatial, geomName = "geom", ...)

\S4method{st_isvalid}{dbSpatial}(dbSpatial, geomName = "geom", ...)
}
\arguments{
\item{geomName}{\code{character string}. The geometry column name in the  \code{\link{dbSpatial}}  object. Default: \code{"geom"}.}

\item{...}{additional arguments passed to methods}

\item{\code{\link{dbSpatial}}}{object}
}
\value{
boolean column vector in database
}
\description{
This function returns whether the specified geometry column in the specified
\code{\link{dbSpatial}}  object is valid or not.
}
\section{Functions}{
\itemize{
\item \code{st_isvalid(dbSpatial)}: Method for \code{dbSpatial} object

}}
\examples{
# Create a data.frame with x and y coordinates and attributes
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a SpatVector from the data.frame
dummy_spatvector <- terra::vect(dummy_data, geom = c("x", "y"))

# Create a duckdb connection
duckdb_conn = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = duckdb_conn,
                      name = "spatVector_proxy",
                      value = dummy_spatvector,
                      overwrite = TRUE)

# Check if geometries are valid
st_isvalid(dbSpatial = db_points)
}
\seealso{
Other geom_scalar: 
\code{\link{st_geometrytype}()}
}
\concept{geom_scalar}
