% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_Read.R
\name{ST_Read}
\alias{ST_Read}
\title{Read spatial data from a file and create a table in a duckdb database}
\usage{
ST_Read(conn, name, value, overwrite = FALSE, ...)
}
\arguments{
\item{conn}{a duckdb connection}

\item{name}{name of the table to be created}

\item{value}{a data.frame or a file path}

\item{overwrite}{logical; if TRUE, overwrite the table if it already exists}

\item{...}{additional arguments to pass to ST_Read}
}
\value{
tbl_dbi
}
\description{
Read spatial data from a file and create a table in a duckdb database
}
\details{
For list of files supported see the documentation below.
\url{https://DuckDB.org/docs/extensions/spatial.html#st_read---read-spatial-value-from-files}
}
\examples{
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)
 
points <- ST_Read(conn = con, name = "points", value = dummy_data, overwrite = TRUE)

points
}
\keyword{geo_construction}
