% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_XMax.R
\name{ST_XMax}
\alias{ST_XMax}
\title{Get maximum x coordinate}
\usage{
ST_XMax(tbl, geomName = "geom")
}
\arguments{
\item{tbl}{name of a table in a duckdb database}

\item{geomName}{name of the column containing the geometry value in the tbl
default = "geom".}
}
\value{
maximum x coordinate
}
\description{
Get maximum x coordinate
}
\examples{
# Create a data.frame with x and y coordinates and attributes
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a duckdb connection
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = con,
                      value = dummy_data,
                      x_colName = "x",
                      y_colName = "y",
                      name = "foo",
                      overwrite = TRUE)
                      
ST_XMax(tbl = db_points)
}
\keyword{spatial_prop}
