% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{tail}
\alias{tail}
\title{tail method for dbSpatial}
\usage{
\S4method{tail}{dbSpatial}(x, n = 6L, ...)
}
\description{
tail method for dbSpatial
}
\seealso{
Other dbData: 
\code{\link{dbData}},
\code{\link{head}()}
}
\concept{dbData}
