% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial.R
\name{ST_Extent}
\alias{ST_Extent}
\title{Get extent of a geometry}
\usage{
ST_Extent(tbl)
}
\arguments{
\item{tbl}{name of a table in a duckdb database}
}
\value{
data.frame of extent of geom column in tbl
}
\description{
Get extent of a geometry
}
\examples{

# Create a data.frame with x and y coordinates and attributes
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a SpatVector from the data.frame
dummy_spatvector <- vect(dummy_data, geom = c("x", "y"))

# Create a duckdb connection
duckdb_conn = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = createDbSpatPoints(conn = duckdb_conn,
                               tbl = "spatVector_proxy",
                               data = dummy_spatvector,
                               overwrite = TRUE)
# Get extent of the table
ext(db_points)
}
\keyword{geo_construction}
