% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_Extent.R
\name{ST_Extent}
\alias{ST_Extent}
\title{Get extent of a geometry}
\usage{
ST_Extent(tbl, geomName = "geom")
}
\arguments{
\item{tbl}{name of a table in a duckdb database containing geometry column}

\item{geomName}{name of the column containing the geometry value in the tbl}
}
\value{
data.frame of extent of geom column in tbl
}
\description{
Get extent of a geometry
}
\examples{
# Create a data.frame with x and y coordinates and attributes
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a duckdb connection
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = con,
                      value = dummy_data,
                      x_colName = "x",
                      y_colName = "y",
                      name = "foo",
                      overwrite = TRUE)
                      
# Get extent of the table
ST_Extent(db_points)
}
\keyword{geo_construction}
