% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_intersects.R
\name{st_intersects}
\alias{st_intersects}
\title{Determine if \code{geometries} between two dbSpatial objects are intersecting}
\usage{
st_intersects(
  g1,
  g1_geom_colName = "geom",
  g1_cols_keep = "all",
  g2,
  g2_geom_colName = "geom",
  g2_cols_keep = "all",
  overwrite = FALSE,
  output_tblName = "intersect_geom"
)
}
\arguments{
\item{g1}{A \code{\link{dbSpatial}} object.}

\item{g1_geom_colName}{\code{character}. The \code{geometry} column name in \code{g1}.
default: "geom"}

\item{g1_cols_keep}{\code{character vector}. The column names in \code{g1} to keep
after the operation. default: 'all' for all columns}

\item{g2}{\code{\link{dbSpatial}} object.}

\item{g2_geom_colName}{\code{character}. The \code{geometry} column name in \code{g2}.
default: "geom"}

\item{g2_cols_keep}{\code{character vector}. The column names in \code{g2} to keep
after the operation. default: 'all' for all columns}

\item{overwrite}{\code{logical}. If TRUE, overwrite an existing table with the
same \code{output_tblName}. default: FALSE}

\item{output_tblName}{\code{character}. The name of the table to store the
resulting data. default: "intersect_geom"}
}
\value{
tbl_dbi
}
\description{
\url{https://postgis.net/docs/ST_Intersects.html}
}
\examples{
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A1", "B1", "C1"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)
 
points <- dbSpatial(conn = con,
                    name = "points", 
                    value = dummy_data, 
                    overwrite = TRUE, 
                    x_colName = "x", 
                    y_colName = "y")

# preview                     
points |> 
  dplyr::mutate(geom_text = ST_AsText(geom))
                    
# Create a second set of points, with B1 and C1 translated by + 100
dummy_data2 <- dummy_data
dummy_data2[c(2,3),c(1,2)] <- dummy_data2[c(2,3),c(1,2)] + 150
dummy_data2$name <- c('A2', 'B2', 'C2')

points2 <- dbSpatial(conn = con,
                    name = "points2", 
                    value = dummy_data2, 
                    overwrite = TRUE, 
                    x_colName = "x", 
                    y_colName = "y")
# preview                     
points2 |> 
  dplyr::mutate(geom_text = ST_AsText(geom))

res <- st_intersects(g1 = points, 
                     g1_cols_keep = c("name"), 
                     g2 = points2,
                     overwrite = TRUE)

res
}
\seealso{
Other spatial_relationships: 
\code{\link{st_contains}()},
\code{\link{st_coveredby}()},
\code{\link{st_covers}()},
\code{\link{st_crosses}()},
\code{\link{st_difference}()},
\code{\link{st_disjoint}()},
\code{\link{st_distance}()},
\code{\link{st_equals}()},
\code{\link{st_touches}()}
}
\concept{spatial_relationships}
