% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_IsValid.R
\name{ST_IsValid}
\alias{ST_IsValid}
\title{Determine if geometry is valid}
\usage{
ST_IsValid(tbl, limit = 10)
}
\arguments{
\item{tbl}{name of a table in a duckdb database}

\item{limit}{maximum number of invalid geometries to return}
}
\value{
boolean vector of whether each geometry in tbl is valid
}
\description{
Determine if geometry is valid
}
\examples{
# Create a data.frame with x and y coordinates and attributes
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a SpatVector from the data.frame
dummy_spatvector <- terra::vect(dummy_data, geom = c("x", "y"))

# Create a duckdb connection
duckdb_conn = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = duckdb_conn,
                      name = "spatVector_proxy",
                      value = dummy_spatvector,
                      overwrite = TRUE)
                      
# Get extent of the table
ST_IsValid(db_points)
}
\keyword{spatial_prop}
