% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_isvalid.R
\name{st_isvalid}
\alias{st_isvalid}
\title{Determine if geometry is valid}
\usage{
st_isvalid(tbl, geomName = "geom")
}
\arguments{
\item{tbl}{name of a table in a duckdb database}

\item{geomName}{name of the column containing the geometry value in the tbl.
default = "geom".}
}
\value{
boolean vector in database
}
\description{
This function returns whether the specified geometry column in the specified
table is valid or not.
}
\examples{
# Create a data.frame with x and y coordinates and attributes
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a SpatVector from the data.frame
dummy_spatvector <- terra::vect(dummy_data, geom = c("x", "y"))

# Create a duckdb connection
duckdb_conn = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = duckdb_conn,
                      name = "spatVector_proxy",
                      value = dummy_spatvector,
                      overwrite = TRUE)
                      
# Get extent of the table
st_isvalid(db_points)
}
\keyword{geom_solo}
