% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_read.R
\name{st_read}
\alias{st_read}
\title{Read spatial data from a file and create a table in a duckdb database}
\usage{
st_read(conn, name, value, overwrite = FALSE, ...)
}
\arguments{
\item{conn}{a duckdb connection}

\item{name}{name of the table to be created}

\item{value}{a data.frame or a file path}

\item{overwrite}{logical; if TRUE, overwrite the table if it already exists}

\item{...}{additional arguments to pass to st_read}
}
\value{
tbl_dbi
}
\description{
Read spatial data from a file and create a table in a duckdb database
}
\details{
For list of files supported see the documentation below.
\url{https://DuckDB.org/docs/extensions/spatial.html#st_read---read-spatial-value-from-files}
}
\examples{
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

write.csv(dummy_data, "dummy_data.csv", row.names = FALSE)
 
data <- st_read(conn = con, 
                name = "points", 
                value = "dummy_data.csv", 
                overwrite = TRUE)

data
}
\keyword{geom_construction}
