% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial.R
\name{ST_Translate}
\alias{ST_Translate}
\title{Translate x, y coordinates by delta x, delta y}
\usage{
ST_Translate(tbl, dx, dy)
}
\arguments{
\item{tbl}{name of a table in a duckdb database}

\item{dx}{delta x}

\item{dy}{delta y}
}
\value{
a duckdb table with translated geometries
}
\description{
Translate x, y coordinates by delta x, delta y
}
\examples{
TODO
}
\keyword{geo_construction}
