% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_Translate.R
\name{ST_Translate}
\alias{ST_Translate}
\title{Translate x, y coordinates by delta x, delta y}
\usage{
ST_Translate(tbl, geomName = "geom", dx, dy)
}
\arguments{
\item{tbl}{name of a table in a duckdb database}

\item{geomName}{name of the column containing the geometry value in the tbl.
default = "geom".}

\item{dx}{delta x}

\item{dy}{delta y}
}
\value{
a duckdb table with translated geometries
}
\description{
Translate x, y coordinates by delta x, delta y
}
\examples{
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)
 
points <- ST_Read(conn = con, name = "points", value = dummy_data, overwrite = TRUE)

points |> 
  dplyr::mutate(geom_text = ST_AsText(geom))

points_translated <- ST_Translate(tbl = points, dx = 100, dy = -20)

points_translated |> 
  dplyr::mutate(geom_text = ST_AsText(geom))
}
\keyword{geo_construction}
