% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial.R
\name{ST_XMax}
\alias{ST_XMax}
\title{Get maximum x coordinate}
\usage{
ST_XMax(tbl)
}
\arguments{
\item{tbl}{name of a table in a duckdb database}
}
\value{
maximum x coordinate
}
\description{
Get maximum x coordinate
}
\examples{
TODO
}
\keyword{spatial_prop}
