% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_xmax.R
\name{st_xmax}
\alias{st_xmax}
\title{Get maximum x coordinate}
\usage{
st_xmax(tbl, geomName = "geom")
}
\arguments{
\item{tbl}{name of a table in a duckdb database}

\item{geomName}{name of the column containing the geometry value in the tbl
default = "geom".}
}
\value{
numerical vector in database
}
\description{
This function returns the maximum x coordinate in each \code{geometry} in
the specified \code{\link{dbSpatial}} object.
}
\examples{
# Create a data.frame with x and y coordinates and attributes
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a duckdb connection
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = con,
                      value = dummy_data,
                      x_colName = "x",
                      y_colName = "y",
                      name = "foo",
                      overwrite = TRUE)
                      
st_xmax(tbl = db_points)
}
\seealso{
Other geom_solo: 
\code{\link{st_geometrytype}()},
\code{\link{st_isvalid}()},
\code{\link{st_ymax}()}
}
\concept{geom_solo}
