% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial.R
\name{ST_YMax}
\alias{ST_YMax}
\title{Get maximum y coordinate}
\usage{
ST_YMax(tbl)
}
\arguments{
\item{tbl}{name of a table in a duckdb database}
}
\value{
maximum y coordinate
}
\description{
Get maximum y coordinate
}
\examples{
TODO
}
\keyword{spatial_prop}
