% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/tessellate.R
\name{tessellate}
\alias{tessellate}
\alias{tessellate,dbSpatial-method}
\title{Tessellate a \code{\link{dbSpatial}}  object}
\usage{
tessellate(
  dbSpatial,
  geomName = "geom",
  name = "tessellation",
  shape = c("hexagon", "square"),
  shape_size = NULL,
  gap = 0,
  radius = NULL,
  overwrite = FALSE,
  ...
)

\S4method{tessellate}{dbSpatial}(
  dbSpatial,
  geomName = "geom",
  name = "tessellation",
  shape = c("hexagon", "square"),
  shape_size = NULL,
  gap = 0,
  radius = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{geomName}{\code{character string}. The geometry column name in the  \code{\link{dbSpatial}}  object. Default: \code{"geom"}.}

\item{name}{\code{character string} name of table to add to \code{\link{dbSpatial}} object. Default: "tessellation".}

\item{shape}{\code{character string}. A character string indicating the shape of the tessellation.
Options are "hexagon" or "square".}

\item{shape_size}{\code{numeric}. the size of the shape in the tessellation.
If \code{NULL}, a default size is calculated. See \code{GiottoClass::tessellate} for details.}

\item{gap}{\code{numeric}. Value indicating the gap between tessellation shapes. Defaults to 0.}

\item{radius}{\code{numeric}. Value specifying the radius for hexagonal tessellation.
This parameter is ignored for square tessellations.}

\item{overwrite}{\code{logical}. Boolean value indicating whether to overwrite an
existing tessellation with the same name. Default: \code{FALSE}.}

\item{...}{Additional arguments passed to methods.}

\item{\code{\link{dbSpatial}}}{object}
}
\value{
\code{\link{dbSpatial}} object
}
\description{
Creates a tessellation on the extent of \code{\link{dbSpatial}}  with specified parameters.
}
\section{Functions}{
\itemize{
\item \code{tessellate(dbSpatial)}: Method for \code{dbSpatial} object

}}
\examples{
coordinates <- data.frame(x = c(100, 200, 300), y = c(500, 600, 700))
attributes <- data.frame(id = 1:3, name = c("A", "B", "C"))

# Combine the coordinates and attributes
dummy_data <- cbind(coordinates, attributes)

# Create a duckdb connection
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")

# Create a duckdb table with spatial points
db_points = dbSpatial(conn = con,
                      value = dummy_data,
                      x_colName = "x",
                      y_colName = "y",
                      name = "foo",
                      overwrite = TRUE)
                      
tessellate(db_points, name = "my_tessellation", shape = "hexagon", shape_size = 60)
}
\seealso{
Other geom_construction: 
\code{\link{st_translate}()}
}
\concept{geom_construction}
